Make Cython-based implementation default
Legacy implementation is still accessible via cuda.parallel.experimental.algorithms.legacy,
and cuda.parallel.experimental.iterator_legacy

Legacy test suite is still accessible via tests_legacy/

Both `pytest tests` and `pytest tests_legacy` pass.
