+++ python/cuda_parallel/cuda/parallel/experimental/algorithms/__init__.py
@@ -3,12 +3,12 @@
 #
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-from .merge_sort import merge_sort as merge_sort
-from .reduce import reduce_into as reduce_into
-from .scan import exclusive_scan as exclusive_scan
-from .scan import inclusive_scan as inclusive_scan
-from .segmented_reduce import segmented_reduce
-from .unique_by_key import unique_by_key as unique_by_key
+from ._cy_merge_sort import merge_sort as merge_sort
+from ._cy_reduce import reduce_into as reduce_into
+from ._cy_scan import exclusive_scan as exclusive_scan
+from ._cy_scan import inclusive_scan as inclusive_scan
+from ._cy_segmented_reduce import segmented_reduce
+from ._cy_unique_by_key import unique_by_key as unique_by_key
 
 __all__ = [
     "merge_sort",
