+++ python/cuda_parallel/cuda/parallel/experimental/algorithms/_cy_scan.py
@@ -16,7 +16,7 @@ from .. import _cy_bindings as cyb
 from .._caching import CachableFunction, cache_with_key
 from .._cccl_for_cy import call_build
 from .._utils import protocols
-from ..cy_iterators._cy_iterators import IteratorBase
+from ..iterators._iterators import IteratorBase
 from ..typing import DeviceArrayLike, GpuStruct
