+++ python/cuda_parallel/cuda/parallel/experimental/algorithms/_legacy_scan.py
@@ -16,23 +16,19 @@ from .. import _cccl as cccl
 from .._bindings import call_build, get_bindings
 from .._caching import CachableFunction, cache_with_key
 from .._utils import protocols
-from ..iterators._iterators import IteratorBase
+from ..iterators_legacy._iterators import IteratorBase
 from ..typing import DeviceArrayLike, GpuStruct
 
 
-def _dtype_validation(dt1, dt2):
-    if dt1 != dt2:
-        raise TypeError(f"dtype mismatch: __init__={dt1}, __call__={dt2}")
-
-
-class _Reduce:
+class _Scan:
     # TODO: constructor shouldn't require concrete `d_in`, `d_out`:
     def __init__(
         self,
         d_in: DeviceArrayLike | IteratorBase,
-        d_out: DeviceArrayLike,
+        d_out: DeviceArrayLike | IteratorBase,
         op: Callable,
         h_init: np.ndarray | GpuStruct,
+        force_inclusive: bool,
     ):
         # Referenced from __del__:
         self.build_result = None
@@ -46,18 +42,25 @@ class _Reduce:
             value_type = numba.typeof(h_init)
         sig = (value_type, value_type)
         self.op_wrapper = cccl.to_cccl_op(op, sig)
-        self.build_result = cccl.DeviceReduceBuildResult()
+        self.build_result = cccl.DeviceScanBuildResult()
         self.bindings = get_bindings()
         error = call_build(
-            self.bindings.cccl_device_reduce_build,
+            self.bindings.cccl_device_scan_build,
             ctypes.byref(self.build_result),
             self.d_in_cccl,
             self.d_out_cccl,
             self.op_wrapper,
-            self.h_init_cccl,
+            cccl.to_cccl_value(h_init),
+            ctypes.c_bool(force_inclusive),
+        )
+
+        self.device_scan = (
+            self.bindings.cccl_device_inclusive_scan
+            if force_inclusive
+            else self.bindings.cccl_device_exclusive_scan
         )
         if error != enums.CUDA_SUCCESS:
-            raise ValueError("Error building reduce")
+            raise ValueError("Error building scan")
 
     def __call__(
         self,
@@ -83,7 +86,7 @@ class _Reduce:
             temp_storage_bytes = ctypes.c_size_t(temp_storage.nbytes)
             d_temp_storage = protocols.get_data_pointer(temp_storage)
 
-        error = self.bindings.cccl_device_reduce(
+        error = self.device_scan(
             self.build_result,
             ctypes.c_void_p(d_temp_storage),
             ctypes.byref(temp_storage_bytes),
@@ -103,19 +106,21 @@ class _Reduce:
     def __del__(self):
         if self.build_result is None:
             return
-        self.bindings.cccl_device_reduce_cleanup(ctypes.byref(self.build_result))
+        self.bindings.cccl_device_scan_cleanup(ctypes.byref(self.build_result))
 
 
 def make_cache_key(
     d_in: DeviceArrayLike | IteratorBase,
-    d_out: DeviceArrayLike,
+    d_out: DeviceArrayLike | IteratorBase,
     op: Callable,
     h_init: np.ndarray,
 ):
     d_in_key = (
         d_in.kind if isinstance(d_in, IteratorBase) else protocols.get_dtype(d_in)
     )
-    d_out_key = protocols.get_dtype(d_out)
+    d_out_key = (
+        d_out.kind if isinstance(d_out, IteratorBase) else protocols.get_dtype(d_out)
+    )
     op_key = CachableFunction(op)
     h_init_key = h_init.dtype
     return (d_in_key, d_out_key, op_key, h_init_key)
@@ -124,30 +129,62 @@ def make_cache_key(
 # TODO Figure out `sum` without operator and initial value
 # TODO Accept stream
 @cache_with_key(make_cache_key)
-def reduce_into(
+def exclusive_scan(
     d_in: DeviceArrayLike | IteratorBase,
-    d_out: DeviceArrayLike,
+    d_out: DeviceArrayLike | IteratorBase,
     op: Callable,
     h_init: np.ndarray,
 ):
-    """Computes a device-wide reduction using the specified binary ``op`` and initial value ``init``.
+    """Computes a device-wide scan using the specified binary ``op`` and initial value ``init``.
 
     Example:
-        Below, ``reduce_into`` is used to compute the minimum value of a sequence of integers.
+        Below, ``scan`` is used to compute an exclusive scan of a sequence of integers.
 
-        .. literalinclude:: ../../python/cuda_parallel/tests/test_reduce_api.py
-            :language: python
-            :dedent:
-            :start-after: example-begin reduce-min
-            :end-before: example-end reduce-min
+        .. literalinclude:: ../../python/cuda_parallel/tests/test_scan_api.py
+          :language: python
+          :dedent:
+          :start-after: example-begin exclusive-scan-max
+          :end-before: example-end exclusive-scan-max
 
     Args:
         d_in: Device array or iterator containing the input sequence of data items
-        d_out: Device array (of size 1) that will store the result of the reduction
+        d_out: Device array that will store the result of the scan
         op: Callable representing the binary operator to apply
-        init: Numpy array storing initial value of the reduction
+        init: Numpy array storing initial value of the scan
 
     Returns:
-        A callable object that can be used to perform the reduction
+        A callable object that can be used to perform the scan
     """
-    return _Reduce(d_in, d_out, op, h_init)
+    return _Scan(d_in, d_out, op, h_init, False)
+
+
+# TODO Figure out `sum` without operator and initial value
+# TODO Accept stream
+@cache_with_key(make_cache_key)
+def inclusive_scan(
+    d_in: DeviceArrayLike | IteratorBase,
+    d_out: DeviceArrayLike | IteratorBase,
+    op: Callable,
+    h_init: np.ndarray,
+):
+    """Computes a device-wide scan using the specified binary ``op`` and initial value ``init``.
+
+    Example:
+        Below, ``scan`` is used to compute an inclusive scan of a sequence of integers.
+
+        .. literalinclude:: ../../python/cuda_parallel/tests/test_scan_api.py
+          :language: python
+          :dedent:
+          :start-after: example-begin inclusive-scan-add
+          :end-before: example-end inclusive-scan-add
+
+    Args:
+        d_in: Device array or iterator containing the input sequence of data items
+        d_out: Device array that will store the result of the scan
+        op: Callable representing the binary operator to apply
+        init: Numpy array storing initial value of the scan
+
+    Returns:
+        A callable object that can be used to perform the scan
+    """
+    return _Scan(d_in, d_out, op, h_init, True)
