+++ python/cuda_parallel/cuda/parallel/experimental/algorithms/legacy.py
@@ -1,0 +1,20 @@
+# Copyright (c) 2024, NVIDIA CORPORATION & AFFILIATES. ALL RIGHTS RESERVED.
+#
+#
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+from ._legacy_merge_sort import merge_sort as merge_sort
+from ._legacy_reduce import reduce_into as reduce_into
+from ._legacy_scan import exclusive_scan as exclusive_scan
+from ._legacy_scan import inclusive_scan as inclusive_scan
+from ._legacy_segmented_reduce import segmented_reduce
+from ._legacy_unique_by_key import unique_by_key as unique_by_key
+
+__all__ = [
+    "merge_sort",
+    "reduce_into",
+    "exclusive_scan",
+    "inclusive_scan",
+    "segmented_reduce",
+    "unique_by_key",
+]
