+++ python/cuda_parallel/cuda/parallel/experimental/iterators/_iterators.py
@@ -20,6 +20,7 @@ from .._utils.protocols import (
     get_shape,
 )
 from ..typing import DeviceArrayLike
+from .._cy_bindings import IteratorState
 
 _DEVICE_POINTER_SIZE = 8
 _DEVICE_POINTER_BITWIDTH = _DEVICE_POINTER_SIZE * 8
@@ -71,8 +72,9 @@ class IteratorBase:
 
     def __init__(
         self,
-        cvalue: ctypes.c_void_p,
+        cvalue,
         numba_type: types.Type,
+        state_type: types.Type,
         value_type: types.Type,
         prefix: str = "",
     ):
@@ -84,6 +86,8 @@ class IteratorBase:
         numba_type
           A numba type representing the type of the input to the advance
           and dereference functions.
+        state_type
+          A numba type of the iterator state.
         value_type
           The numba type of the value returned by the dereference operation.
         prefix
@@ -91,20 +95,24 @@ class IteratorBase:
         """
         self.cvalue = cvalue
         self.numba_type = numba_type
+        self.state_type = state_type
         self.value_type = value_type
         self.prefix = prefix
+        self.kind_ = self.__class__.iterator_kind_type(self.value_type)
+        self.state_ = IteratorState(self.cvalue)
 
     @property
     def kind(self):
-        return self.__class__.iterator_kind_type(self.value_type)
+        return self.kind_
 
     # TODO: should we cache this? Current docs environment doesn't allow
     # using Python > 3.7. We could use a hand-rolled cached_property if
     # needed.
     @property
     def ltoirs(self) -> Dict[str, bytes]:
-        advance_abi_name = f"{self.prefix}advance_" + _get_abi_suffix(self.kind)
-        deref_abi_name = f"{self.prefix}dereference_" + _get_abi_suffix(self.kind)
+        abi_suffix = _get_abi_suffix(self.kind)
+        advance_abi_name = f"{self.prefix}advance_{abi_suffix}"
+        deref_abi_name = f"{self.prefix}dereference_{abi_suffix}"
         advance_ltoir, _ = cached_compile(
             self.__class__.advance,
             (
@@ -124,8 +132,8 @@ class IteratorBase:
         return {advance_abi_name: advance_ltoir, deref_abi_name: deref_ltoir}
 
     @property
-    def state(self) -> ctypes.c_void_p:
-        return ctypes.cast(ctypes.pointer(self.cvalue), ctypes.c_void_p)
+    def state(self) -> IteratorState:
+        return self.state_
 
     @staticmethod
     def advance(state, distance):
@@ -178,10 +186,12 @@ class RawPointer(IteratorBase):
 
     def __init__(self, ptr: int, value_type: types.Type):
         cvalue = ctypes.c_void_p(ptr)
-        numba_type = types.CPointer(types.CPointer(value_type))
+        state_type = types.CPointer(value_type)
+        numba_type = types.CPointer(state_type)
         super().__init__(
             cvalue=cvalue,
             numba_type=numba_type,
+            state_type=state_type,
             value_type=value_type,
         )
 
@@ -229,9 +239,14 @@ class CacheModifiedPointer(IteratorBase):
     def __init__(self, ptr: int, ntype: types.Type, prefix: str):
         cvalue = ctypes.c_void_p(ptr)
         value_type = ntype
-        numba_type = types.CPointer(types.CPointer(value_type))
+        state_type = types.CPointer(value_type)
+        numba_type = types.CPointer(state_type)
         super().__init__(
-            cvalue=cvalue, numba_type=numba_type, value_type=value_type, prefix=prefix
+            cvalue=cvalue,
+            numba_type=numba_type,
+            state_type=state_type,
+            value_type=value_type,
+            prefix=prefix,
         )
 
     @staticmethod
@@ -253,10 +268,12 @@ class ConstantIterator(IteratorBase):
     def __init__(self, value: np.number):
         value_type = numba.from_dtype(value.dtype)
         cvalue = to_ctypes(value_type)(value)
-        numba_type = types.CPointer(value_type)
+        state_type = value_type
+        numba_type = types.CPointer(state_type)
         super().__init__(
             cvalue=cvalue,
             numba_type=numba_type,
+            state_type=state_type,
             value_type=value_type,
         )
 
@@ -279,10 +296,12 @@ class CountingIterator(IteratorBase):
     def __init__(self, value: np.number):
         value_type = numba.from_dtype(value.dtype)
         cvalue = to_ctypes(value_type)(value)
-        numba_type = types.CPointer(value_type)
+        state_type = value_type
+        numba_type = types.CPointer(state_type)
         super().__init__(
             cvalue=cvalue,
             numba_type=numba_type,
+            state_type=state_type,
             value_type=value_type,
         )
 
@@ -352,6 +371,7 @@ def make_transform_iterator(it, op: Callable):
         def __init__(self, it: IteratorBase, op: CUDADispatcher):
             self._it = it
             self._op = CachableFunction(op.py_func)
+            state_type = it.state_type
             numba_type = it.numba_type
             # TODO: it would be nice to not need to compile `op` to get
             # its return type, but there's nothing in the numba API
@@ -366,6 +386,7 @@ def make_transform_iterator(it, op: Callable):
             super().__init__(
                 cvalue=it.cvalue,
                 numba_type=numba_type,
+                state_type=state_type,
                 value_type=value_type,
             )
 
@@ -396,12 +417,11 @@ def make_advanced_iterator(it: IteratorBase, /, *, offset: int = 1):
 
         def __init__(self, it: IteratorBase, advance_steps: int):
             self._it = it
-            cvalue_advanced = to_ctypes(it.value_type)(
-                it.cvalue + it.value_type(advance_steps)
-            )
+            cvalue_advanced = type(it.cvalue)(it.cvalue.value + advance_steps)
             super().__init__(
                 cvalue=cvalue_advanced,
                 numba_type=it.numba_type,
+                state_type=it.state_type,
                 value_type=it.value_type,
             )
