+++ python/cuda_parallel/tests_legacy/test_unique_by_key_api.py
@@ -8,7 +8,7 @@ def test_unique_by_key():
     import cupy as cp
     import numpy as np
 
-    import cuda.parallel.experimental.algorithms._cy_unique_by_key as algorithms
+    import cuda.parallel.experimental.algorithms.legacy as algorithms
 
     def compare_op(lhs, rhs):
         return np.uint8(lhs == rhs)
